pass

+++ internal/wasm/interpreter/interpreter.go
@@ -1389,7 +1389,7 @@ func (ce *callEngine) callNativeFunc(ctx context.Context, callCtx *wasm.CallCont
 								panic(wasmruntime.ErrRuntimeIntegerOverflow)
 							}
 						}
-						ce.pushValue(uint64(int32(v)))
+						ce.pushValue(uint64(uint32(int32(v))))
 					case wazeroir.SignedInt64:
 						v := math.Trunc(float64(math.Float32frombits(uint32(ce.popValue()))))
 						res := int64(v)
